refactoring observables for settings and playlists, improvements
