+++ app/src/main/res/values/dimen.xml
@@ -30,7 +30,7 @@
     <dimen name="close_handle_icon_padding">10dp</dimen>
     <dimen name="main_floating_button_size">95dp</dimen>
 
-    <dimen name="albumDetail_chip_radius">20dp</dimen>
+    <dimen name="albumDetail_chip_radius">10dp</dimen>
     <dimen name="albumDetail_chip_elevation">2dp</dimen>
     <dimen name="albumDetail_title_fontSize">24sp</dimen>
     <dimen name="albumDetailScreen_top_padding">5dp</dimen>
